+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## [unreleased]
 
+### Added
+
+- Type annotation for the rust part of the codebase.
+- `can_primary_expire` new argument to `create_key` function call.
 
 ## [0.10.0] - 2022-09-20